Fix GH workflow to exit successfully
- In the `check-hooks-pr` workflow we use
`continue-on-error: true` because we
don't want to show the workflow as
failed if one step exits with 1.
- In the `tag-version` workflow we split
into two jobs, one with `continue-on-error: false`
(default) because we want the workflow to show
as failed if the tests fail and one with
`continue-on-error: true` where we don't care if
it fails when checking the version.